dynamic header links + route page created
